完善腾讯云存储的单元测试

+++ Magicodes.Storage/Magicodes.Storage.Tests/QiNiuStorageTests.cs
@@ -1,31 +1,53 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
+﻿// ======================================================================
+//   
+//           Copyright (C) 2018-2020 湖南心莱信息科技有限公司    
+//           All rights reserved
+//   
+//           filename : QiNiuStorageTests.cs
+//           description :
+//   
+//           created by 雪雁 at  2018-06-07 10:31
+//           Mail: wenqiang.li@xin-lai.com
+//           QQ群：85318032（技术交流）
+//           Blog：http://www.cnblogs.com/codelove/
+//           GitHub：https://github.com/xin-lai
+//           Home：http://xin-lai.com
+//   
+// ======================================================================
+
+using System;
+using System.IO;
 using System.Threading.Tasks;
-using Magicodes.Storage.QiNiu.Core;
-using Xunit;
-using Shouldly;
 using Magicodes.Storage.Core;
+using Magicodes.Storage.QiNiu.Core;
 using Qiniu.Storage;
 using Qiniu.Util;
-using System.IO;
+using Shouldly;
+using Xunit;
 
 namespace Magicodes.Storage.Tests
 {
     [Trait("Group", "七牛云存储测试")]
     public class QiNiuStorageTests : TestBase, IDisposable
     {
-        QiNiuStorageProvider qiNiuStorageProvider;
-        string localPath;
+        private readonly string localPath;
+        private readonly QiNiuStorageProvider qiNiuStorageProvider;
+
         public QiNiuStorageTests()
         {
             localPath = Path.Combine(Directory.GetCurrentDirectory(), "Files");
             if (!Directory.Exists(localPath)) Directory.CreateDirectory(localPath);
 
             //七牛云注册即有免费空间,请替换为自己的
-            qiNiuStorageProvider = new QiNiuStorageProvider("33Ag3A3lSRrZ3VHyaCravN61g5DWhz6C24I5Zp_r", "qbepDnLi6Y4b4DH9XchxP-qTKddZeA5iLyZitE-U");
+            qiNiuStorageProvider = new QiNiuStorageProvider("33Ag3A3lSRrZ3VHyaCravN61g5DWhz6C24I5Zp_r",
+                "qbepDnLi6Y4b4DH9XchxP-qTKddZeA5iLyZitE-U");
 
             StorageProvider = qiNiuStorageProvider;
+        }
+
+        public void Dispose()
+        {
+            //TODO:清理数据
         }
 
         [Theory(DisplayName = "文件删除测试")]
@@ -38,7 +60,7 @@ namespace Magicodes.Storage.Tests
             bucketsResult.Result.Contains(containerName).ShouldBe(true);
 
             await Assert.ThrowsAnyAsync<StorageException>(async () =>
-               await StorageProvider.DeleteBlob(containerName, blobName)
+                await StorageProvider.DeleteBlob(containerName, blobName)
             );
 
             var putPolicy = new PutPolicy
@@ -53,18 +75,8 @@ namespace Magicodes.Storage.Tests
 
             var testFilePath = Path.Combine(localPath, blobName);
             File.WriteAllText(testFilePath, GetTestContent());
-            var result = um.UploadFile(testFilePath, blobName, token, new PutExtra()
-            {
-
-            });
-            if (result.Code == 200)
-            {
-                await StorageProvider.DeleteBlob(containerName, blobName);
-            }
-        }
-
-        public void Dispose() {
-            //TODO:清理数据
+            var result = um.UploadFile(testFilePath, blobName, token, new PutExtra());
+            if (result.Code == 200) await StorageProvider.DeleteBlob(containerName, blobName);
         }
     }
 }
